Wrap ostree_repo_resolve_rev and ostree_repo_pull_with_options
Add small wappers around those to ostree functions.

OstreeAsyncProgress got a minimal wrapper so it can be used in the
function signatures (but really can only be used as a nil value)

ResolveRev is fully functional and a direct mirror of the C function,
for PullWithOptions rather then passing an options GVariant a
PullOptions go struct is defined which the wrapper converts to a
GVariant to both work around the fact that GVariant isn't usable from Go
and to make the API slightly nicer. For now only only the
override-remote-name and refs options are support as those were the ones
i happened to need in my usage.

Signed-off-by: Sjoerd Simons <sjoerd.simons@collabora.co.uk>